requirements.txt
flask
python-dotenv
groq
sentence-transformers
chromadb
PyMuPDF
python-docx
langchain-text-splitters
